backend/requirements.txt
asgiref==3.11.1
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
dj-database-url==2.2.0
gunicorn==21.2.0
Pillow==11.2.1
psycopg2-binary==2.9.10
sqlparse==0.5.5
whitenoise==6.8.2
